envvault-ml/requirements.txt
flask==3.1.1
flask-cors==5.0.1
scikit-learn==1.6.1
pandas==2.2.3
numpy<2
joblib==1.4.2
python-dotenv==1.1.0
gunicorn<23
mlflow==2.13.0
